Default tags and multiple series per measurement
The configured plugs (if supplied in a config.json file) are now grouped
into a single measurement and identified by their relevant tags to stop
measurement sprall in Influx. The series are also easier to identify as
the device alias is polled from the plug and included as a tag.

Relay state and Wifi RSSI have been added to the default measurements as
they are useful for debugging.